refactor(ui): move AppIcon to dbflux_components::icons; embedded_bytes free fn (S3) (#102)
